entirely remove Kiali community operator
The Kiali community operator is no longer being updated. It is now obsolete and should no longer be published in the community catalog.

For those users in the OKD community that still wish to use the Kiali Operator can install the operator via the Kiali operator helm chart.

For those Red Hat subscribers using Kiali, they still have access to the Kiali Operator published in the Red Hat catalog - that is still actively being managed and updated.

Signed-off-by: John Mazzitelli <mazz@redhat.com>